+++ packages/remix-deno/index.ts
@@ -15,6 +15,7 @@ export {
 
 export {
   createSession,
+  defer,
   isCookie,
   isSession,
   json,
@@ -63,8 +64,11 @@ export type {
   SessionIdStorageStrategy,
   SessionStorage,
   SignFunction,
+  TypedDeferredData,
   TypedResponse,
   UnsignFunction,
   UploadHandler,
   UploadHandlerPart,
+  V2_HtmlMetaDescriptor,
+  V2_MetaFunction,
 } from "@remix-run/server-runtime";