+++ packages/remix-cloudflare-pages/CHANGELOG.md
@@ -1,5 +1,47 @@
 # `@remix-run/cloudflare-pages`
 
+## 1.11.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.11.0`
+
+## 1.10.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.10.1`
+
+## 1.10.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.10.0`
+
+## 1.9.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.9.0`
+
+## 1.8.2
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.8.2`
+
+## 1.8.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.8.1`
+
 ## 1.8.0
 
 ### Patch Changes