+++ packages/remix-netlify/CHANGELOG.md
@@ -1,5 +1,48 @@
 # `@remix-run/netlify`
 
+## 1.11.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.11.0`
+
+## 1.10.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.10.1`
+
+## 1.10.0
+
+### Patch Changes
+
+- Improve performance of `isBinaryType` ([#4761](https://github.com/remix-run/remix/pull/4761))
+- Updated dependencies:
+  - `@remix-run/node@1.10.0`
+
+## 1.9.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.9.0`
+
+## 1.8.2
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.8.2`
+
+## 1.8.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.8.1`
+
 ## 1.8.0
 
 ### Patch Changes