+++ packages/remix-server-runtime/CHANGELOG.md
@@ -1,5 +1,101 @@
 # `@remix-run/server-runtime`
 
+## 1.11.0
+
+### Minor Changes
+
+- Added support for [Vanilla Extract](https://vanilla-extract.style) via the `unstable_vanillaExtract` future flag. **IMPORTANT:** Features marked with `unstable` are … unstable. While we're confident in the use cases they solve, the API and implementation may change without a major version bump. ([#5040](https://github.com/remix-run/remix/pull/5040))
+- Add support for CSS side-effect imports via the `unstable_cssSideEffectImports` future flag. **IMPORTANT:** Features marked with `unstable` are … unstable. While we're confident in the use cases they solve, the API and implementation may change without a major version bump. ([#4919](https://github.com/remix-run/remix/pull/4919))
+- Add support for CSS Modules via the `unstable_cssModules` future flag. **IMPORTANT:** Features marked with `unstable` are … unstable. While we're confident in the use cases they solve, the API and implementation may change without a major version bump. ([#4852](https://github.com/remix-run/remix/pull/4852))
+
+### Patch Changes
+
+- Added the `v2_errorBoundary` future flag to opt into the next version of Remix's `ErrorBoundary` behavior. This removes the separate `CatchBoundary` and `ErrorBoundary` and consolidates them into a single `ErrorBoundary`, following the logic used by `errorElement` in React Router. You can then use `isRouteErrorResponse` to differentiate between thrown `Response`/`Error` instances. ([#4918](https://github.com/remix-run/remix/pull/4918))
+
+  ```jsx
+  // Current (Remix v1 default)
+  import { useCatch } from "@remix-run/react";
+
+  export function CatchBoundary() {
+    let caught = useCatch();
+    return (
+      <p>
+        {caught.status} {caught.data}
+      </p>
+    );
+  }
+
+  export function ErrorBoundary({ error }) {
+    return <p>{error.message}</p>;
+  }
+  ```
+
+  ```jsx
+  // Using future.v2_errorBoundary
+  import { isRouteErrorResponse, useRouteError } from "@remix-run/react";
+
+  export function ErrorBoundary() {
+    let error = useRouteError();
+
+    return isRouteErrorResponse(error) ? (
+      <p>
+        {error.status} {error.data}
+      </p>
+    ) : (
+      <p>{error.message}</p>
+    );
+  }
+  ```
+
+- Introduces the `defer()` API from `@remix-run/router` with support for server-rendering and HTTP streaming. This utility allows you to defer values returned from `loader` functions by returning promises instead of resolved values. This has been refered to as _"sending a promise over the wire"_. ([#4920](https://github.com/remix-run/remix/pull/4920))
+
+  Informational Resources:
+
+  - <https://gist.github.com/jacob-ebey/9bde9546c1aafaa6bc8c242054b1be26>
+  - <https://github.com/remix-run/remix/blob/main/decisions/0004-streaming-apis.md>
+
+  Documentation Resources (better docs specific to Remix are in the works):
+
+  - <https://reactrouter.com/en/main/utils/defer>
+  - <https://reactrouter.com/en/main/components/await>
+  - <https://reactrouter.com/en/main/hooks/use-async-value>
+  - <https://reactrouter.com/en/main/hooks/use-async-error>
+
+## 1.10.1
+
+No significant changes to this package were made in this release. [See the releases page on GitHub](https://github.com/remix-run/remix/releases/tag/remix%401.10.1) for an overview of all changes in v1.10.1.
+
+## 1.10.0
+
+### Minor Changes
+
+- Update Remix to use new data APIs introduced in React Router v6.4 ([#4900](https://github.com/remix-run/remix/pull/4900))
+
+### Patch Changes
+
+- Export `V2_HtmlMetaDescriptor` and `V2_MetaFunction` types from runtime packages ([#4943](https://github.com/remix-run/remix/pull/4943))
+- Fix `V2_MetaFunction` to return `V2_HtmlMetaDescriptor[]` type ([#4947](https://github.com/remix-run/remix/pull/4947))
+
+## 1.9.0
+
+### Patch Changes
+
+- Fix `TypedResponse` so that Typescript correctly shows errors for incompatible types in `loader` and `action` functions. ([#4734](https://github.com/remix-run/remix/pull/4734))
+- Fix error boundary tracking for multiple errors bubbling to the same boundary ([#4829](https://github.com/remix-run/remix/pull/4829))
+- Fixed an issue where a loader's `Request` object reflected `method: "POST"` on document submissions ([`a74e51830`](https://github.com/remix-run/remix/commit/a74e51830ec7ecb3ad30e45013270ebf71d7b425))
+
+## 1.8.2
+
+### Patch Changes
+
+- Remove `instanceof Response` checks in favor of `isResponse` ([#4782](https://github.com/remix-run/remix/pull/4782))
+- Fix performance regression with creation of `@remix-run/router` static handler ([#4790](https://github.com/remix-run/remix/pull/4790))
+- Update dependency for `@remix-run/router` to `v1.0.5` ([`bd84a9317`](https://github.com/remix-run/remix/commit/bd84a931770a6b5e20c2f21839b4322023432b25))
+
+## 1.8.1
+
+No significant changes to this package were made in this release. [See the releases page on GitHub](https://github.com/remix-run/remix/releases/tag/remix%401.8.1) for an overview of all changes in v1.8.1.
+
 ## 1.8.0
 
 ### Minor Changes